backend/ML/Recommendation/requirement.txt
google-cloud-bigquery
google-cloud-storage
flask
flask-cors
